soucis du mac mini resolu en passant les double en float, apparement la version OpenCL des dump n'a pas de probleme avec les double ce qui n'est pas le cas du OpenCL/mmini

+++ burning_ship_kernel.cl
@@ -1,9 +1,9 @@
-kernel void		burning_ship(global double *env, global int *couleur)
+kernel void		burning_ship(global float *env, global int *couleur)
 {
 	int			pixel;
-	double2		c;
-	double2		z;
-	double		tmp;
+	float2		c;
+	float2		z;
+	float		tmp;
 	int			i;
 
 	pixel = get_global_id(0);
